Rename nosetests execution scripts and vars to unit_tests (#infra)
Better name in general. Not referring nose framework which we don't use now
anymore.

Remove nosetests_root.sh because the tests it's pointing does not exists.

This is breaking change but it shouldn't be such a pain because I don't think
there are many users (other than from terminal) who use the
`UNIT_TESTS_ARGS` Makefile var or the `tests-unit-only` Makefile target.

`UNIT_TESTS_ARGS` now pass arguments to the `unittest discover` python module.
Change documentation to reflect that.

(cherry picked from commit df1c352ab66aba63d380d0d5f912d8a329b08dd6)
