+++ tests/unit_tests/pyanaconda_tests/modules/payloads/test_module_payloads.py
@@ -21,8 +21,8 @@ from unittest import TestCase
 from unittest.mock import patch, create_autospec, DEFAULT
 
 from pyanaconda.core.constants import SOURCE_TYPE_LIVE_OS_IMAGE
-from tests.nosetests.pyanaconda_tests import patch_dbus_publish_object, check_dbus_object_creation
-from tests.nosetests.pyanaconda_tests.modules.payloads.payload.module_payload_shared import PayloadKickstartSharedTest
+from tests.unit_tests.pyanaconda_tests import patch_dbus_publish_object, check_dbus_object_creation
+from tests.unit_tests.pyanaconda_tests.modules.payloads.payload.module_payload_shared import PayloadKickstartSharedTest
 
 from pyanaconda.modules.common.containers import PayloadContainer
 from pyanaconda.modules.common.errors.payload import SourceSetupError, SourceTearDownError
